feat(TicTacToe): add action listener to each JButton
The program will have several buttons that players can press in order to select a space to mark either an x or an o. This commit adds an action listener to all 9 buttons in our 3x3 grid so that something can happen once they are pressed. In this case, that something will be marking the space as either and x or an o.