fix(TicTacToe): fix game title not displaying
The game title 'Tic-Tac-Toe' was not displaying before indicating a player's turn. Not with the use of a try-catch block it displays the title for 2 seconds, then displays which player's turn it is. I also had to move the firstTurn() method to be at the bottom of the constructor after the setVisible() method for this to work properly.