feat(TicTacToe): Add actionPerformed() method body
The actionPerformed() method will allow players to add X and O characaters to the game board by clicking on specific buttons. Once a player clicks a spot, if that spot is empty, then the code will set the text to that spot to be the current player's character. Finally, once a player selects their empty spot, the textField of the game will switch its text to say who the new current player is.