fix(TicTacToe): Match look and feel of UI with look and feel cross-platform
There was an issue where JButtons in the application were not able to have their color changed in a way that matched the aesthetic of the system's look and feel. A try-catch block was added at the top of top of the constructor that gets the cross-platform look and feel and configure's the game's UI to have the same look and feel. Now JButtons can have their colors changed accordingly.