fix(TicTacToe): Fix tie feature
The logic of the tie feature would trigger when all spots are marked, but it should only trigger if a winner was not already declared. Sometimes there would be a winner but the winning condition was met by also selecting a spot that would fill up the board, triggering the textField to say tie when it should have declared a winner. To fix this, I changed the conditions that check for a winner from being all if statements to also being if else statements